enums_refactor linting cleanup
