+++ src/control_panel.py
@@ -1,5 +1,5 @@
-from constants import Location, move_elevations
-from game_states import GameStates
+from constants import move_elevations
+from enums import Location, GameStates
 from utilities import get_cone_target_hexes_at_location
