+++ src/ui.py
@@ -1,5 +1,4 @@
-from constants import message_count, block_size, view_port, tile_size
-from render_functions import game_font, margin
+from constants import message_count, block_size, view_port, tile_size, game_font, margin
 
 
 class DisplayInfo:
